% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgit.R
\name{dbgit_init}
\alias{dbgit_init}
\title{Initialize dbgit reposiory}
\usage{
dbgit_init(wksp_folder)
}
\arguments{
\item{wksp_folder}{indicates the absolute location of databricks workspace e.g. /Users/some@email.com for pointing to a home folder}
}
\description{
This function allow to configure a folder to be used as sync repository with databricks
}
