% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgit.R
\name{dbgit_push}
\alias{dbgit_push}
\title{Push a directory to databricks}
\usage{
dbgit_push(overwrite = FALSE)
}
\arguments{
\item{overwrite}{is a boolean indicating if we should overwrite the file if already exists. By default is FALSE.}
}
\description{
This function allow you to sync pushing a local git folder TO databricks
}
